Rename googleAPI's file name
